@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
     <RootNamespace>AssemblyToProcess</RootNamespace>
     <SignAssembly>true</SignAssembly>
     <AssemblyOriginatorKeyFile>CoreLibSigned.snk</AssemblyOriginatorKeyFile>
@@ -9,12 +9,12 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="2.2.0" />
+    <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="5.0.0" />
   </ItemGroup>
 
   <ItemGroup>
     <Reference Include="AutoDI">
-      <HintPath>..\..\NuGet\AutoDI\lib\netstandard2.0\AutoDI.dll</HintPath>
+      <HintPath>..\..\..\NuGet\AutoDI\lib\netstandard2.0\AutoDI.dll</HintPath>
     </Reference>
   </ItemGroup>
 </Project>
