@@ -10,19 +10,19 @@ namespace AutoDI.AssemblyGenerator
     {
         public static object GetStaticProperty<TContainingType>(this Assembly assembly, string propertyName, Type containerType = null) where TContainingType : class
         {
-            if (assembly == null) throw new ArgumentNullException(nameof(assembly));
-            if (propertyName == null) throw new ArgumentNullException(nameof(propertyName));
+            if (assembly is null) throw new ArgumentNullException(nameof(assembly));
+            if (propertyName is null) throw new ArgumentNullException(nameof(propertyName));
 
             string typeName = TypeMixins.GetTypeName(typeof(TContainingType), containerType);
             Type type = assembly.GetType(typeName);
-            if (type == null)
+            if (type is null)
                 throw new AssemblyGetPropertyException($"Could not find '{typeof(TContainingType).FullName}' in '{assembly.FullName}'");
 
             PropertyInfo property = type.GetProperty(propertyName);
-            if (property == null)
+            if (property is null)
                 throw new AssemblyGetPropertyException($"Could not find property '{propertyName}' on type '{type.FullName}'");
 
-            if (property.GetMethod == null)
+            if (property.GetMethod is null)
                 throw new AssemblyGetPropertyException($"Property '{type.FullName}.{propertyName}' does not have a getter");
 
             if (!property.GetMethod.IsStatic)
@@ -33,15 +33,15 @@ namespace AutoDI.AssemblyGenerator
 
         public static object InvokeStatic<TContainingType>(this Assembly assembly, string methodName, params object[] parameters) where TContainingType : class
         {
-            if (assembly == null) throw new ArgumentNullException(nameof(assembly));
-            if (methodName == null) throw new ArgumentNullException(nameof(methodName));
+            if (assembly is null) throw new ArgumentNullException(nameof(assembly));
+            if (methodName is null) throw new ArgumentNullException(nameof(methodName));
 
             Type type = assembly.GetType(typeof(TContainingType).FullName);
-            if (type == null)
+            if (type is null)
                 throw new AssemblyInvocationException($"Could not find '{typeof(TContainingType).FullName}' in '{assembly.FullName}'");
 
             MethodInfo method = type.GetMethod(methodName);
-            if (method == null)
+            if (method is null)
                 throw new AssemblyInvocationException($"Could not find method '{methodName}' on type '{type.FullName}'");
 
             if (!method.IsStatic)
@@ -52,7 +52,7 @@ namespace AutoDI.AssemblyGenerator
 
         public static void InvokeEntryPoint(this Assembly assembly)
         {
-            if (assembly == null) throw new ArgumentNullException(nameof(assembly));
+            if (assembly is null) throw new ArgumentNullException(nameof(assembly));
             
             assembly.EntryPoint.Invoke(null, new object[assembly.EntryPoint.GetParameters().Length]);
         }
@@ -60,7 +60,7 @@ namespace AutoDI.AssemblyGenerator
         public static object InvokeGeneric<TGeneric>(this Assembly assembly, object target, string methodName, params object[] parameters)
         {
             Type genericType = assembly.GetType(typeof(TGeneric).FullName);
-            if (genericType == null)
+            if (genericType is null)
                 throw new AssemblyInvocationException($"Could not find generic parameter type '{typeof(TGeneric).FullName}' in '{assembly.FullName}'");
 
             return InvokeGeneric(assembly, genericType, target, methodName, parameters);
@@ -68,10 +68,10 @@ namespace AutoDI.AssemblyGenerator
 
         public static object InvokeGeneric(this Assembly assembly, Type genericType, object target, string methodName, params object[] parameters)
         {
-            if (assembly == null) throw new ArgumentNullException(nameof(assembly));
-            if (genericType == null) throw new ArgumentNullException(nameof(genericType));
-            if (target == null) throw new ArgumentNullException(nameof(target));
-            if (methodName == null) throw new ArgumentNullException(nameof(methodName));
+            if (assembly is null) throw new ArgumentNullException(nameof(assembly));
+            if (genericType is null) throw new ArgumentNullException(nameof(genericType));
+            if (target is null) throw new ArgumentNullException(nameof(target));
+            if (methodName is null) throw new ArgumentNullException(nameof(methodName));
 
             Type targetType = target.GetType();
 
@@ -80,7 +80,7 @@ namespace AutoDI.AssemblyGenerator
                     .SelectMany(@interface => @interface.GetRuntimeMethods().Where(x => x.Name == methodName)));
 
             MethodInfo method = methods.FirstOrDefault(m => !m.IsStatic && m.IsGenericMethodDefinition);
-            if (method == null)
+            if (method is null)
                 throw new AssemblyInvocationException($"Could not find method '{methodName}' on type '{targetType.FullName}'");
             
             MethodInfo genericMethod = method.MakeGenericMethod(genericType);
@@ -90,11 +90,11 @@ namespace AutoDI.AssemblyGenerator
 
         public static object CreateInstance<T>(this Assembly assembly, Type containerType = null)
         {
-            if (assembly == null) throw new ArgumentNullException(nameof(assembly));
+            if (assembly is null) throw new ArgumentNullException(nameof(assembly));
 
             string typeName = TypeMixins.GetTypeName(typeof(T), containerType);
             Type type = assembly.GetType(typeName);
-            if (type == null)
+            if (type is null)
                 throw new AssemblyCreateInstanceException($"Could not find '{typeName}' in '{assembly.FullName}'");
 
             foreach (ConstructorInfo ctor in type.GetConstructors().OrderBy(c => c.GetParameters().Length))
@@ -110,17 +110,17 @@ namespace AutoDI.AssemblyGenerator
 
         public static object Resolve<T>(this Assembly assembly, Type containerType = null)
         {
-            if (assembly == null) throw new ArgumentNullException(nameof(assembly));
+            if (assembly is null) throw new ArgumentNullException(nameof(assembly));
 
             string typeName = TypeMixins.GetTypeName(typeof(T), containerType);
             Type type = assembly.GetType(typeName);
-            if (type == null)
+            if (type is null)
                 throw new AssemblyCreateInstanceException($"Could not find '{typeName}' in '{assembly.FullName}'");
 
 
             IServiceProvider provider = DI.GetGlobalServiceProvider(assembly);
 
-            if (provider == null)
+            if (provider is null)
             {
                 //TODO: Better exception
                 throw new AssemblyInvocationException($"Could not find service provider for '{assembly.FullName}'");
