@@ -26,7 +26,7 @@ namespace AutoDI.Build
             ConstructorInfo baseCtor = (baseType ?? typeof(object))
                 .GetConstructor(BindingFlags.Instance | BindingFlags.Public | BindingFlags.NonPublic, null, new Type[0],
                     null);
-            if (baseCtor == null)
+            if (baseCtor is null)
                 throw new Exception($"Could not find constructor for '{(baseType ?? typeof(object)).FullName}'");
             processor.Emit(OpCodes.Call, module.ImportReference(baseCtor));
             processor.Emit(OpCodes.Ret);
@@ -44,8 +44,8 @@ namespace AutoDI.Build
 
         public static FieldDefinition CreateStaticReadonlyField<T>(this ModuleDefinition moduleDefintion, string name, bool @public)
         {
-            if (moduleDefintion == null) throw new ArgumentNullException(nameof(moduleDefintion));
-            if (name == null) throw new ArgumentNullException(nameof(name));
+            if (moduleDefintion is null) throw new ArgumentNullException(nameof(moduleDefintion));
+            if (name is null) throw new ArgumentNullException(nameof(name));
 
             return moduleDefintion.CreateStaticReadonlyField(name, @public, moduleDefintion.Get<T>());
         }
@@ -53,8 +53,8 @@ namespace AutoDI.Build
         public  static FieldDefinition CreateStaticReadonlyField(this ModuleDefinition moduleDefinition, string name, 
             bool @public, TypeReference type)
         {
-            if (moduleDefinition == null) throw new ArgumentNullException(nameof(moduleDefinition));
-            if (name == null) throw new ArgumentNullException(nameof(name));
+            if (moduleDefinition is null) throw new ArgumentNullException(nameof(moduleDefinition));
+            if (name is null) throw new ArgumentNullException(nameof(name));
 
             return new FieldDefinition(name,
                 (@public ? FieldAttributes.Public : FieldAttributes.Private) | FieldAttributes.Static |
@@ -63,8 +63,8 @@ namespace AutoDI.Build
 
         public static FieldDefinition CreateReadonlyField<T>(this ModuleDefinition moduleDefinition, string name, bool @public)
         {
-            if (moduleDefinition == null) throw new ArgumentNullException(nameof(moduleDefinition));
-            if (name == null) throw new ArgumentNullException(nameof(name));
+            if (moduleDefinition is null) throw new ArgumentNullException(nameof(moduleDefinition));
+            if (name is null) throw new ArgumentNullException(nameof(name));
 
             return moduleDefinition.CreateReadonlyField(name, @public, moduleDefinition.Get<T>());
         }
@@ -72,8 +72,8 @@ namespace AutoDI.Build
         public static FieldDefinition CreateReadonlyField(this ModuleDefinition moduleDefinition,
             string name, bool @public, TypeReference type)
         {
-            if (moduleDefinition == null) throw new ArgumentNullException(nameof(moduleDefinition));
-            if (name == null) throw new ArgumentNullException(nameof(name));
+            if (moduleDefinition is null) throw new ArgumentNullException(nameof(moduleDefinition));
+            if (name is null) throw new ArgumentNullException(nameof(name));
 
             return new FieldDefinition(name,
                 (@public ? FieldAttributes.Public : FieldAttributes.Private) |
@@ -87,10 +87,10 @@ namespace AutoDI.Build
 
         public static MethodReference GetDefaultConstructor(this ModuleDefinition moduleDefinition, Type targetType)
         {
-            if (moduleDefinition == null) throw new ArgumentNullException(nameof(moduleDefinition));
+            if (moduleDefinition is null) throw new ArgumentNullException(nameof(moduleDefinition));
 
             var defualtCtor = targetType.GetConstructor(new Type[0]);
-            if (defualtCtor == null)
+            if (defualtCtor is null)
             {
                 throw new InvalidOperationException($"Could not find default constructor for '{targetType.FullName}'");
             }
@@ -104,7 +104,7 @@ namespace AutoDI.Build
 
         public static MethodReference GetConstructor(this ModuleDefinition moduleDefinition, Type targetType, params Type[] argumentTypes)
         {
-            if (moduleDefinition == null) throw new ArgumentNullException(nameof(moduleDefinition));
+            if (moduleDefinition is null) throw new ArgumentNullException(nameof(moduleDefinition));
 
             var ctors = targetType.GetConstructors();
 
@@ -134,12 +134,12 @@ namespace AutoDI.Build
         public static MethodReference GetMethod(this ModuleDefinition moduleDefinition,
             Type containingType, string methodName)
         {
-            if (moduleDefinition == null) throw new ArgumentNullException(nameof(moduleDefinition));
-            if (containingType == null) throw new ArgumentNullException(nameof(containingType));
-            if (methodName == null) throw new ArgumentNullException(nameof(methodName));
+            if (moduleDefinition is null) throw new ArgumentNullException(nameof(moduleDefinition));
+            if (containingType is null) throw new ArgumentNullException(nameof(containingType));
+            if (methodName is null) throw new ArgumentNullException(nameof(methodName));
 
             MethodInfo method = containingType.GetMethod(methodName);
-            if (method == null) throw new InvalidOperationException($"Could not find method '{methodName}' on '{containingType.FullName}'");
+            if (method is null) throw new InvalidOperationException($"Could not find method '{methodName}' on '{containingType.FullName}'");
 
             return moduleDefinition.ImportReference(method);
         }
